Move context hooks into their own directory
