+++ apps/frontend/src/library/hooks/context/useSavedWordsContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import SavedWordsContext from "../context/SavedWordsContext";
+import SavedWordsContext from "../../context/SavedWordsContext";
 
 const useSavedWordsContext = () => {
   const context = useContext(SavedWordsContext);
