+++ apps/frontend/src/library/hooks/context/useThemeContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import ThemeContext from "../context/ThemeContext";
+import ThemeContext from "../../context/ThemeContext";
 
 const useThemeContext = () => {
   const context = useContext(ThemeContext);
