+++ apps/frontend/src/library/hooks/context/useUserContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import UserContext from "../context/UserContext";
+import UserContext from "../../context/UserContext";
 
 const useUserContext = () => {
   const context = useContext(UserContext);
